train_csm/pyproject.toml
[project]
name = "train-csm"
version = "0.1.0"
description = "Training script for Sesame CSM using Unsloth"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "unsloth @ git+https://github.com/unslothai/unsloth.git",
    "unsloth_zoo",
    "transformers>=4.48.0",
    "datasets>=3.2.0",
    "torch>=2.1.0",
    "accelerate>=0.26.0",
    "scipy",
    "soundfile",
    "librosa",
    "wandb",
    "python-dotenv",
    "PyYAML"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
